+++ CHANGELOG.md
@@ -5,7 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-## [1.1.1] - 2020-06-26
+## [1.1.2] - 2022-06-29
+### Added
+- A Generic L-System
+- Generic Generator and some implementations
+- An interface for PlayerPrefs and an implementation that allows to separate data by scope
+
+
+## [1.1.1] - 2022-06-26
 ### Added
 - Approximately methods in FloatExtensions.cs